# openapi.yaml
openapi: "3.1.0"
info:
  title: FlagBridge API
  version: "1.0.0"
  description: |
    Feature flags for the whole team — AI-powered. Open-core platform.

    **Editions:** Community (CE) and Pro. Pro endpoints require a valid license key
    and are gated by internal `pro.*` flags (ADR-001).

    **Base URL:** `https://api.flagbridge.io`

    **Authentication:**
    - `Bearer <JWT>` — admin routes (obtained from `/v1/auth/login`)
    - `Bearer <SDK_KEY>` — SDK/evaluation routes (`fb_sk_eval_*`)
servers:
  - url: https://api.flagbridge.io
    description: Production
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SDKKey:
      type: http
      scheme: bearer
      description: SDK API key (fb_sk_eval_* / fb_sk_mgmt_* / fb_sk_test_*)
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
                          name:
                            type: string
                          role:
                            type: string
        "401":
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # PROJECTS
  # ============================================================
  /v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List all projects
      security:
        - BearerJWT: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      security:
        - BearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/projects/{slug}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a project by slug
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        "404":
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # ENVIRONMENTS
  # ============================================================
  /v1/projects/{slug}/environments:
    get:
      tags: [Environments]
      operationId: listEnvironments
      summary: List environments for a project
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
      responses:
        "200":
          description: Environments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Environment'
    post:
      tags: [Environments]
      operationId: createEnvironment
      summary: Create an environment
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
                color:
                  type: string
                sort_order:
                  type: integer
      responses:
        "201":
          description: Created environment
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Environment'

  # ============================================================
  # FLAGS
  # ============================================================
  /v1/projects/{slug}/flags:
    get:
      tags: [Flags]
      operationId: listFlags
      summary: List flags for a project
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tag filter
      responses:
        "200":
          description: Flags
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Flag'
    post:
      tags: [Flags]
      operationId: createFlag
      summary: Create a flag
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, name, type]
              properties:
                key:
                  type: string
                name:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum: [boolean, string, number, json]
                default_value: {}
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Created flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Flag'
  /v1/projects/{slug}/flags/{key}:
    get:
      tags: [Flags]
      operationId: getFlag
      summary: Get a flag by key
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        "200":
          description: Flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Flag'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Flags]
      operationId: updateFlag
      summary: Update a flag
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                default_value: {}
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Updated flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Flag'
    delete:
      tags: [Flags]
      operationId: deleteFlag
      summary: Delete a flag
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        "204":
          description: Deleted

  # ============================================================
  # FLAG STATES (per environment)
  # ============================================================
  /v1/projects/{slug}/flags/{key}/states/{env}:
    get:
      tags: [Flag States]
      operationId: getFlagState
      summary: Get flag state for an environment
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
        - $ref: '#/components/parameters/EnvSlug'
      responses:
        "200":
          description: Flag state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FlagState'
    put:
      tags: [Flag States]
      operationId: setFlagState
      summary: Set flag state for an environment
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
        - $ref: '#/components/parameters/EnvSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                value: {}
                rollout:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: Updated state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FlagState'

  # ============================================================
  # TARGETING RULES
  # ============================================================
  /v1/projects/{slug}/flags/{key}/rules/{env}:
    get:
      tags: [Targeting Rules]
      operationId: getTargetingRules
      summary: Get targeting rules for a flag in an environment
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
        - $ref: '#/components/parameters/EnvSlug'
      responses:
        "200":
          description: Targeting rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TargetingRule'
    put:
      tags: [Targeting Rules]
      operationId: setTargetingRules
      summary: Replace targeting rules for a flag in an environment
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/FlagKey'
        - $ref: '#/components/parameters/EnvSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                rules:
                  type: array
                  items:
                    type: object
                    required: [name, priority, conditions, value, enabled]
                    properties:
                      name:
                        type: string
                      priority:
                        type: integer
                      conditions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Condition'
                      value: {}
                      enabled:
                        type: boolean
      responses:
        "200":
          description: Updated rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TargetingRule'

  # ============================================================
  # EVALUATION
  # ============================================================
  /v1/evaluate:
    post:
      tags: [Evaluation]
      operationId: evaluateFlag
      summary: Evaluate a single flag
      security:
        - SDKKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flag_key, project, environment]
              properties:
                flag_key:
                  type: string
                project:
                  type: string
                environment:
                  type: string
                context:
                  $ref: '#/components/schemas/EvalContext'
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EvalResult'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/evaluate/batch:
    post:
      tags: [Evaluation]
      operationId: batchEvaluate
      summary: Evaluate multiple flags in one request
      security:
        - SDKKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, environment, flag_keys]
              properties:
                project:
                  type: string
                environment:
                  type: string
                flag_keys:
                  type: array
                  items:
                    type: string
                context:
                  $ref: '#/components/schemas/EvalContext'
      responses:
        "200":
          description: Map of flag_key to evaluation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      flags:
                        type: object
                        additionalProperties:
                          $ref: '#/components/schemas/EvalResult'

  # ============================================================
  # SSE
  # ============================================================
  /v1/sse/{environment}:
    get:
      tags: [Evaluation]
      operationId: sseStream
      summary: Server-Sent Events stream for real-time flag updates
      security:
        - SDKKey: []
      parameters:
        - name: environment
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  # ============================================================
  # TESTING API (CE basic / Pro full)
  # ============================================================
  /v1/testing/sessions:
    post:
      tags: [Testing]
      operationId: createTestingSession
      summary: Create a testing session
      security:
        - SDKKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, environment]
              properties:
                project:
                  type: string
                environment:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TestingSession'
    get:
      tags: [Testing]
      operationId: listTestingSessions
      summary: List testing sessions (Pro)
      security:
        - BearerJWT: []
      x-flagbridge-edition: pro
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestingSession'
  /v1/testing/sessions/{id}:
    delete:
      tags: [Testing]
      operationId: deleteTestingSession
      summary: Delete a testing session
      security:
        - SDKKey: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        "204":
          description: Deleted
  /v1/testing/sessions/{id}/overrides:
    get:
      tags: [Testing]
      operationId: getTestingOverrides
      summary: Get all overrides in a testing session
      security:
        - SDKKey: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        "200":
          description: Overrides
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestingOverride'
    put:
      tags: [Testing]
      operationId: setTestingOverrides
      summary: Set (replace all) overrides for a testing session
      security:
        - SDKKey: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [overrides]
              properties:
                overrides:
                  type: array
                  items:
                    $ref: '#/components/schemas/TestingOverride'
      responses:
        "200":
          description: Updated overrides
  /v1/testing/sessions/{id}/overrides/{flag}:
    delete:
      tags: [Testing]
      operationId: deleteTestingOverride
      summary: Remove a single flag override from a testing session
      security:
        - SDKKey: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
        - name: flag
          in: path
          required: true
          schema:
            type: string
          description: Flag key
      responses:
        "204":
          description: Deleted
...
  /v1/projects/{slug}/webhooks/{webhook_id}/test:
    post:
      tags: [Webhooks]
      operationId: testWebhook
      summary: Send a test delivery to a webhook
      security:
        - BearerJWT: []
      parameters:
        - $ref: '#/components/parameters/ProjectSlug'
        - $ref: '#/components/parameters/WebhookID'
      responses:
        "200":
          description: Test delivery sent

  # ============================================================
  # API KEYS
  # ============================================================
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...